==> src/03_Sprint4_FINAL/00_DataUnderstanding.ipynb <==
# Purpose of the Notebook:
The purpose of this notebook is to view and become familiar with the data set. What are the columns, how many columns are there and what are the data types of each column. 

# Data Understanding

# imports
import pandas as pd

# read in the dataframe
url = 'https://raw.githubusercontent.com/joa24jm/21-12-13_air-qual/master/data/01_raw/AirQualityUCI.csv'
df = pd.read_csv(url, sep=';', decimal=',')

## Fist inspection

# first output
df.head(10)


# number of rows and columns
df.shape

## Rows and columns

#first row
df.head(1)

#last row
df.tail(1)

#first and last column
df.iloc[:, [0, -1]]

## Informations

df.describe()

# number of rows and columns
df.shape

# checking datatypes of all columns
df.dtypes

> While familiarizing ourselves with the data set, we quickly noticed some superfluous and incorrect formatting. The features "Unnamed: 15" and "Unnamed: 16" are superfluous, since they contain absolutely no value and they have no proper name. We have recognized a wrong formatting in the features "Date" and "Time", because they are not formatted as data type "date" or "time". Furthermore, we recognized that the value -200 occurs relatively often and must therefore be examined more closely.

## Local copy of dataframe 

df.to_csv('../../data/csv/AirQuality/00_AirQuality_original.csv', sep=';', index = False)
