==> src/03_Sprint4_FINAL/01_DataPreparation.ipynb <==
# Purpose of the Notebook: DataPreparation
The purpose of this notebook is to clean up the data set. This is done by removing superfluous columns and rows that, for example, consist only of nan values and do not add any value to the dataset. The date and time is adjusted and converted to a datetime type. In addition, the feature names are changed to a more conventional form, which is easier to work with.

# imports
import pandas as pd
import X1_DataPreparation as data_prep
import orga_functions as org

# read in the dataframe
path = org.path("00_AirQuality_original.csv")
df = pd.read_csv(path, sep=';')

## Cleaning the data 

Remove columns that only contains NaN/None/Null-values and therefore do not add value to the dataframe.

# Remove
df.dropna(axis = "columns", how = "all", inplace = True)

df

<br/><br/>

## Correction of the date

Call the generated method, which was created for the correction and formatting of the date and time. 

# Methode: correct datetime 
df = data_prep.df_correct_datetime(df)

df

Check if the formatting of the type into a datetime was successful.

df.dtypes

## Customization of column names

Call the generated method created for the correction and formatting of the column names. 

# Method: customize feature names
df = data_prep.df_change_columns(df)

df.head(7)

<br/><br/>

## Filter and exclude implausible values

Compare the dataframe with the values from the feature list and check for plausibility.

# Method: filter with featureList 
df = data_prep.df_filter_plausible(df, feature_df_path = "02_AlleFeatureList.csv")

df

## Save

Save the clean dataset locally for the further work.

# save 
new_path = org.path("01_AirQuality_processed.csv")
df.to_csv(new_path, sep=';', index = False)
