==> src/03_Sprint4_FINAL/01_DataPreparation_profiling.ipynb <==
# Purpose of the Chapter: DataPreparation profiling
The purpose of this notebook is to create a report using Panda's profiling, which can be saved and viewed as HTML. 

# imports
import pandas as pd
from pandas_profiling import ProfileReport
import orga_functions as org

# read in the dataframes
df = pd.read_csv(org.path("01_AirQuality_processed.csv"), sep=';')
feature_df = pd.read_csv(org.path("00_FeatureList_original.csv"), sep =";")

Create the Pandas Profile Report and give him the name "AirQuality Report". 

profile = ProfileReport(df, title = "AirQuality Report")

df

feature_df

# name with discription
name = feature_df["names"].values
discription= feature_df["erläuterung"].values
name_dict = dict(zip(name, discription))

profile.config.variables.descriptions = name_dict

# save the report as HTML
file_path = org.path("AirQuality_report.html")
profile.to_file(output_file = file_path)
