==> src/03_Sprint4_FINAL/06_Feature_Engineering.ipynb <==
# Purpose of the Notebook

In this notebook, an attempt is made to improve the performance of the model by doing feature engineering. Feature engineering is the process of using domain knowledge to extract features (characteristics, properties, attributes) from raw data.

# Feature Engineering

# imports
import pandas as pd
import X3_Forecasting as forecast
import orga_functions as org
import datetime

# read in the dataframe
df = pd.read_csv(org.path("02_AirQuality_processed.csv"), sep=';')
#df = pd.read_csv(org.path("03_AirQuality_normalized.csv"), sep=';')
df

# Ensure the timestamp-datetype
df["date"] = pd.to_datetime(df["date"], format = "%Y-%m-%d %H:%M:%S")

## New Features

These are the features which have been thought of and how they were created:

### Weekday

Categorical features 'monday' to 'sunday' which are 1 if true and 0 if false.

# weekday(number) for every date
df["weekday"] = [x.weekday() for x in df.date]

df

# columns for every single weekday (0/1 values)
df["monday"] = df["tuesday"] = df["wednesday"] = df["thursday"] = df["friday"] = df["saturday"] = df["sunday"] = 0

# set weekday for every row
for i in range(len(df)):
    if df.loc[i]["weekday"] == 0:
        df.at[i, "monday"] = 1
        
    elif df.loc[i]["weekday"] == 1:
        df.at[i, "tuesday"] = 1
        
    elif df.loc[i]["weekday"] == 2:
        df.at[i, "wednesday"] = 1
        
    elif df.loc[i]["weekday"] == 3:
        df.at[i, "thursday"] = 1
        
    elif df.loc[i]["weekday"] == 4:
        df.at[i, "friday"] = 1
        
    elif df.loc[i]["weekday"] == 5:
        df.at[i, "saturday"] = 1
        
    elif df.loc[i]["weekday"] == 6:
        df.at[i, "sunday"] = 1

df.drop(columns=["weekday"], inplace =True)

df

### Seasons

Categorical features 'winter' to 'autumn' which are true if the timestamp is within that season.

# source: https://stackoverflow.com/questions/16139306/determine-season-given-timestamp-in-python-using-datetime
from datetime import date, datetime

Y = 2000 # dummy leap year to allow input X-02-29 (leap day)
seasons = [('winter', (date(Y,  1,  1),  date(Y,  3, 20))),
           ('spring', (date(Y,  3, 21),  date(Y,  6, 20))),
           ('summer', (date(Y,  6, 21),  date(Y,  9, 22))),
           ('autumn', (date(Y,  9, 23),  date(Y, 12, 20))),
           ('winter', (date(Y, 12, 21),  date(Y, 12, 31)))]


def get_season(now):
    if isinstance(now, datetime):
        now = now.date()
    now = now.replace(year=Y)
    return next(season for season, (start, end) in seasons
                if start <= now <= end)


# future time - better performance!
df["date"] = df["date"]+ pd.DateOffset(hours=6, minutes=0)

# season name for every date
df["season"] = [get_season(x) for x in df.date]

# check for 4 seasons
print(df["season"].unique())

# columns for every single season (0/1 values)
df["spring"] = df["summer"] = df["autumn"] = df["winter"] = 0

# set season for every row
for i in range(len(df)):
    if df.loc[i]["season"] == "spring":
        df.at[i, "spring"] = 1
        
    elif df.loc[i]["season"] == "summer":
        df.at[i, "summer"] = 1
        
    elif df.loc[i]["season"] == "autumn":
        df.at[i, "autumn"] = 1
        
    elif df.loc[i]["season"] == "winter":
        df.at[i, "winter"] = 1

# drop column season
df.drop(columns = ["season"], inplace = True)

### Daytime

Categorical features which represent the time of day as 'early_morning' to 'night'.

df["early_morning"] = df["mid_morning"] = df["midday"] = df["afternoon"] = df["evening"] = df["night"] = 0

for i in range(len(df)):
    
    curr_cel = df.loc[i]["date"]
    
    if curr_cel.hour >= 6 and curr_cel.hour < 10:
        df.at[i, "early_morning"] = 1
        
    elif curr_cel.hour >= 10 and curr_cel.hour < 12 :
        df.at[i, "mid_morning"] = 1
        
    elif curr_cel.hour >= 12 and curr_cel.hour < 14 :
        df.at[i, "midday"] = 1
    
    elif curr_cel.hour >= 14 and curr_cel.hour < 17 :
        df.at[i, "afternoon"] = 1
    
    elif curr_cel.hour >= 17 and curr_cel.hour < 21 :
        df.at[i, "evening"] = 1
    else:
        df.at[i, "night"] = 1

df

# save dataset locally
new_path = org.path("04_AirQuality_NewFeatures.csv")
df.to_csv(new_path, sep=';', index = False)

## Test

Below you will find the results of the performance tests that were conducted to test whether the new features improve performance or not.

# set date column as index column
df.set_index("date", inplace = True)

# searching for best features by checking the best mean_absolute_error (LinearRegression) (with 6hrs shift)

#0.0633526390377646
features_with_weekday = ['pt08_s1_co',"c6h6_gt","pt08_s2_nmhc","no2_gt", "pt08_s4_no2","t",
                       "monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday", "ah"]


# 0.06070736267619392
# 0.06003062680060758  # hrs +6
features_with_season = ['pt08_s1_co',"c6h6_gt","pt08_s2_nmhc","no2_gt", "pt08_s4_no2","t",
                        "spring", "summer", "autumn", "winter", "ah"]


# 0.062101402957687124
features_weekday_season = ['pt08_s1_co',"c6h6_gt","pt08_s2_nmhc","no2_gt", "pt08_s4_no2","t",
                  "monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday", 
                  "spring", "summer", "autumn", "winter", "ah"]


# 0.06819995689500315)
features_with_daytime = ['pt08_s1_co',"c6h6_gt","pt08_s2_nmhc","no2_gt", "pt08_s4_no2","t",
                        "early_morning", "mid_morning", "midday", "afternoon", "evening","night", "ah"]

#### Final Result

best_features = ['pt08_s1_co',"c6h6_gt","pt08_s2_nmhc","no2_gt", "pt08_s4_no2","t",
                        "spring", "summer", "autumn", "winter", "ah"] 

features = best_features

## Forecasting

Below is the area used for testing the performance.

#### Target

#define target column
target = 'ah_target'

#### Shift

# shift abs humidty values by 6 hrs
df[target] = df.ah.shift(periods=-6)

# drop rows where target is unknown (last 6 rows with NaN ah_target)
df.dropna(subset=[target], inplace=True)

#### Preparation

# Train/ Test-split
from sklearn.model_selection import train_test_split
training, test = train_test_split(df, test_size=1000, shuffle=False, random_state=1999)

# input/ target-split
X_train = training[features]
y_train = training[target]

x_test = test[features]
y_test = test[target]

#### Linear Regression

# training the regression model
from sklearn.linear_model import LinearRegression
clf_lin_reg = LinearRegression()
clf_lin_reg.fit(X_train, y_train)

#### Performance Check

# make prediction
pred_linr_y = clf_lin_reg.predict(x_test)

# calculate mean absolute error
from sklearn.metrics import mean_absolute_error
mean_absolute_error(y_test, pred_linr_y)

# target vs prediction by model
forecast.check_df(x_test, y_test, clf_lin_reg)
