==> src/03_Sprint4_FINAL/07_Forecasting.ipynb <==
# Purpose of the Notebook:

This notebook is all about prediction.
For this purpose, two different regressors are taken into account and their performance is compared.
Furthermore, these regressors are saved for further usage.


# Forecasting

# imports
import pandas as pd
import X3_Forecasting as forecast
import orga_functions as org
import pickle

from sklearn.metrics import mean_absolute_error
from sklearn.metrics import d2_absolute_error_score

# import the regressor
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor 

# read in the data set
#df = pd.read_csv(org.path("02_AirQuality_processed.csv"), sep=';', index_col='date')
df = pd.read_csv(org.path("03_AirQuality_normalized.csv"), sep=';', index_col="date")

> We can distinguish between the data in its original form and the normalised form.
Normalising the AH values leads to the problem that the results can no longer be interpreted by humans.
As an temporary solution, all values except AH were normalised.

df

## Features

# list of features which will be used for the algorithm - based on 05_Feature_selection
features = ["t", "ah"]

## Target

target = 'ah_target'

# shift of the absolute humidty values by 6 hrs
target = 'ah_target'
df[target] = df.ah.shift(periods=-6)

df.head(8)

> As we can see, the ah_target value from the first line equals to the ah value from the seventh line indicating that the shift was successful.

df.shape

## Limitation of the ML_DataFrame

# drop rows where target is unknown
df.dropna(subset=[target], inplace=True)

> due to the shift last 6 rows with NaN ah_target are dropped

df

df.shape

## Preparation Data

The data set is split into a training set and a test set

# import the split function
from sklearn.model_selection import train_test_split

# split
training, test = train_test_split(df, test_size=1000, shuffle=False, random_state=1999)

# classification of features and the target
X_train = training[features]
y_train = training[target]
x_test = test[features]
y_test = test[target]

> Both sets are subdivided into the features and the target, thus all other columns are not included any further

X_train

> As an example, nox was not defined as a feature. It can be seen here that the nox column is no longer a part of the dataframe.

## Models/Regressors

### Linear Regression

# The regressor gets trained 
clf_lin_reg = LinearRegression()
clf_lin_reg.fit(X_train, y_train)

# values are estimated by the regressor
pred_linr_y = clf_lin_reg.predict(x_test)

### Performance Check

# visual comparison of the actual and estimated values
check_linr = forecast.check_df(x_test, y_test, clf_lin_reg)
check_linr

#mean_absolute_error of the predicted AH value compared to the actual future AH value
mean_absolute_error(y_test, pred_linr_y)

# variance of the predicted AH value compared to the actual future AH value
d2_absolute_error_score(y_test, pred_linr_y)

## Decision Tree

# The regressor gets trained
clf_tree = DecisionTreeRegressor(random_state = 0) 
clf_tree.fit(X_train, y_train)

# values are estimated by the regressor
pred_tree_y = clf_tree.predict(x_test)

### Performance Check

# visual comparison of the actual and estimated values
check_tree_df = forecast.check_df(x_test, y_test, clf_tree)
check_tree_df

#mean_absolute_error of the prediction and the actual values
mean_absolute_error(y_test, pred_tree_y)

# variance of the prediction and the actual values
d2_absolute_error_score(y_test, pred_tree_y)

## Saving

### Save Training/Test Split

# save training dataset locally
new_path = org.path("00_training.csv")
training.to_csv(new_path, sep=';', index = True)

# save test dataset locally
new_path = org.path("00_testing.csv")
test.to_csv(new_path, sep=';', index = True)

# save the x_test dataset
x_test_path = org.path("00_x_test.csv")
x_test.to_csv(x_test_path, sep=';', index = True)

# save the y_test dataset
y_test_path = org.path("00_y_test.csv")
y_test.to_csv(y_test_path, sep=';', index = True)

# save the x_train dataset
x_train_path = org.path("00_x_train.csv")
X_train.to_csv(x_train_path, sep=';', index = True)

# save the y_train dataset
y_train_path = org.path("00_y_train.csv")
y_train.to_csv(y_train_path, sep=';', index = True)

### Save Model

# the trained LinearRegressor is saved as a pickle file
new_path = org.path("LinearRegression_ah_regressor.pkl")

with open(new_path, 'wb') as f:
    pickle.dump(clf_lin_reg, f)

# the trained DecisionTree is saved as a pickle file
new_path = org.path("DecisionTree_ah_regressor.pkl")

with open(new_path, 'wb') as f:
    pickle.dump(clf_tree, f)

> These regressors are stored to be used later for unit tests and also in reality.
