==> src/03_Sprint4_FINAL/07_Forecasting_2_test.ipynb <==
# Purpose of the Notebook:

In this notebook, it is verified that the saving of the regressors was successful and that they can be further used.

# Forecasting Test

## Imports

# imports
import pandas as pd
import pickle
import X3_Forecasting as forecast
import orga_functions as org

from sklearn.metrics import mean_absolute_error

# import the features (x_test) and the actual value of the target (y_test)
x_test = pd.read_csv(org.path("00_x_test.csv"), sep=';')
y_test = pd.read_csv(org.path("00_y_test.csv"), sep=';')

x_test.head(10)

# import model
with open(org.path("LinearRegression_ah_regressor.pkl"), 'rb') as f:
    linr_clf = pickle.load(f)

## Result

# predict for x_test
pred_y = forecast.pred_array(x_test, linr_clf)

# visual comparison of the actual and estimated values
check = forecast.check_df(x_test,y_test, linr_clf)
check

mean_absolute_error(y_test, pred_y)
