==> src/03_Sprint4_FINAL/08_Create_LiveData_JSON.ipynb <==
# Purpose of the Notebook:

In this notebook, a sample of simulated live data is generated.
Those samples can be used in later notebooks to check whether the forecasting algorithm works in every possible situation.

# Create JSON LiveData

# import
import json
import pandas as pd
import numpy as np
import orga_functions as org

from X1_DataPreparation import change_column
from X1_DataPreparation import og_column_list
from X1_DataPreparation import og_int_column_list

# read in the feature codebook
path = org.path("02_AlleFeatureList.csv")
feature_df = pd.read_csv(path, sep =";")

## FeatureList DataPreparation

feature_df

### Original Names

# original Names List
original_names = og_column_list()

# replace names by original/raw names
def og_name(name):
    for og in original_names:
        if change_column(og) == name:
            return og
    return name

feature_df.names = [og_name(x) for x in feature_df.names]

feature_df

# the name column is set as the new index column
feature_df.set_index('names', inplace = True)

feature_df[["min", "max"]]

## Create new entries

### Functions

inner functions

# gets the min for the given feature
def get_min(feature_name):
    return feature_df.loc[feature_name]["min"]

# gets the max for the given feature
def get_max(feature_name):
    return feature_df.loc[feature_name]["max"]
                         

main functions

# returns a correct value for the given feature
def get_correct(feat_name):
    if feat_name in og_int_column_list():
        value = np.random.randint(get_min(feat_name), get_max(feat_name))
        return value
    else:
        value = np.random.uniform(get_min(feat_name), get_max(feat_name))
        return round(value, 2)

# returns a to small value for the given feature
def get_too_small(feat_name):
    if feat_name in og_int_column_list():
        value = np.random.randint(-250, get_min(feat_name)-1)
        return value
    else:
        value = np.random.uniform(-250, get_min(feat_name)-1)
        return round(value, 2)

# returns a to big value for the given feature
def get_too_big(feat_name):
    if feat_name in og_int_column_list():
        value = np.random.randint(get_max(feat_name)+1, 5000)
        return value
    else:
        value = np.random.uniform(get_max(feat_name)+1, 5000)
        return round(value, 2)

 ### Entry with only correct values

# generate values for the all columns besides the date and time columns
corr_dict = {}
for i in feature_df.index.values:
    corr_dict[i] = get_correct(i)

corr_dict["Date"] = "07/02/2005"
corr_dict["Time"] = '00.00.00'

### Entry with some too small values

# generate values for the all columns besides the date and time columns
small_dict = {}

change = 1

for i in feature_df.index.values:
    if change %2 == 0:
        small_dict[i] = get_correct(i)
    else:
        small_dict[i] = get_too_small(i)
    change+=1

small_dict["Date"] = "07/02/2005"
small_dict["Time"] = '01.00.00'

### Entry some too big values

# generate values for the all columns besides the date and time columns
big_dict = {}

# variable 
change = 1

for i in feature_df.index.values:
    if change % 2 == 0:
        big_dict[i] = get_correct(i)
    else:
        big_dict[i] = get_too_big(i)
    change+=1

big_dict["Date"] = "07/02/2005"
big_dict["Time"] = '02.00.00'

### Entry with mix values

# generate values for the all columns besides the date and time columns
mix_dict = {}

change = 1

for i in feature_df.index.values:
    if change % 3 == 0:              
        mix_dict[i] = get_too_big(i)
    elif change % 2 == 0:            
        mix_dict[i] = get_too_small(i)
    else:                            
        mix_dict[i] = get_correct(i)
    change+=1

mix_dict["Date"] = "07/02/2005"
mix_dict["Time"] = '03.00.00'

### Entry with missing  values

missing_dict = {}

missing_dict["Date"] = "07/02/2005"
missing_dict["Time"] = '04.00.00'

# generate same values for the all columns besides the date and time columns
missing_dict["CO(GT)"] = get_correct('CO(GT)')
missing_dict["PT08.S1(CO)"] = get_correct('PT08.S1(CO)')
missing_dict["NO2(GT)"] = get_correct('NO2(GT)')

missing_dict["NOx(GT)"] = np.nan

### entry with an unknown column

dict_unknown = {}
for i in feature_df.index.values:
    dict_unknown[i] = get_correct(i)
dict_unknown["unknown"] = "low"
dict_unknown["Date"] = "07/02/2005"
dict_unknown["Time"] = '05.00.00'

### more entries

corr_dict2 = {}
for i in feature_df.index.values:
    corr_dict2[i] = get_correct(i)

corr_dict2["Date"] = "07/02/2005"
corr_dict2["Time"] = '06.00.00'

corr_dict3 = {}
for i in feature_df.index.values:
    corr_dict3[i] = get_correct(i)

corr_dict3["Date"] = "07/02/2005"
corr_dict3["Time"] = '07.00.00'

## Save  LiveData in JSON File

#function that saves the data in a JSON form
def entry_list(*entries):
    list = []
    for e in entries:
        list.append(e)
    return list

### Save

# convert each every into a JSON-like form
first_entry = entry_list(corr_dict)
second_entry = entry_list(small_dict)
third_entry = entry_list(big_dict)
fourth_entry = entry_list(mix_dict)
fifth_entry = entry_list(missing_dict)
sixth_entry = entry_list(dict_unknown)

# convert the last two entries into a single JSON-like form
other_entries = entry_list(corr_dict2, corr_dict3)

# Save as JSON file
new_path = org.path("01_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(first_entry, jf, indent = 2)

# Save as JSON file
new_path = org.path("02_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(second_entry, jf, indent = 2)

# Save as JSON file
new_path = org.path("03_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(third_entry, jf, indent = 2)

# Save as JSON file
new_path = org.path("04_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(fourth_entry, jf, indent = 2)

# Save as JSON file
new_path = org.path("05_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(fifth_entry, jf, indent = 2)

# Save as JSON file
new_path = org.path("06_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(sixth_entry, jf, indent = 2)

# Save as JSON file
new_path = org.path("07_LiveData.json")

with open(new_path, 'w') as jf:
    json.dump(other_entries, jf, indent = 2)

### Exemplarily Output

#Output for the first entry
print(json.dumps(first_entry, indent = 2))

#Output for the two last entries 
print(json.dumps(other_entries, indent = 2))
